Shuffling to better encapsulate building behaviors.
